+++ src/nft/components/details/AssetPriceDetails.tsx
@@ -1,0 +1,2 @@
+
+
